requirements.txt.txt
Django==1.10.3
django-tinymce
pillow